backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
python-socketio>=5.10.0
python-multipart>=0.0.6
loguru>=0.7.2
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
google-generativeai>=0.8.3
tenacity>=8.2.3
pytest>=7.0.0
httpx>=0.27.0
slowapi>=0.1.9
